Get extra subscription info for new push protocol
Firefox 46+ and Chrome 50+ support the new encrypted web push protocol.
We obtain and forward the two parameters 'auth' and 'p256dh', both a
member of the PushSubscription object we obtain from
PushManager.subscribe(), to OneSignal. Firefox 44 & 45 do not support
the 'auth' property so we only forward each property if it exists.
Chrome < 50 does not support either 'auth' or 'p256dh'.