Rename notify button animation to prevent conflict
Our notify button pulse animation is defined as "pulse" which is not
very unique and is shared by animate.css's popular library. On our own
page at https://onesignal.com/webpush, we use the animate.css library
and the naming conflict prevents the animation from being shown. This
commit renames the animation to "notifyButtonPulse" which should prevent
name conflicts from happening. The @keyframes was also added as it looks
like @-webkit-keyframes was added but @keyframes was never added.