Fix return detected Chinese language variant
Our backend supports sending notifications as zh-Hans and zh-Hant as
Chinese (simplified) and Chinese (traditional). A user's language is
set by our SDK which we pull from navigator.language. Previously, a
the first two characters of the user's language tag would be returned,
and so a user with a navigator.language of 'zh-TW', 'zh-Hans', or
'zh-CN' would be forwarded to OneSignal as only 'zh'. We would then
convert users marked only as 'zh' to 'zh-Hans' or Chinese (simplified).
This commit maps popular Chinese language tags or otherwise returns
traditional Chinese as the default Chinese language. Tests added.