Recover from wiped IndexedDB if push signal is sent
If a user has their IndexedDB wiped, but the service worker is still
active, on Chrome and Firefox we will recover the user's ID based on the
subscription endpoint from OneSignal's server and retry receiving the
notification. If the recovery attempt fails, the user is unsubscribed
from push notifications so that he does not receive "New site updates"
messages.