Fix HTTPS modal prompt subscription
The recently added feature where we don't allow multiple subscription
calls broke the HTTPS modal prompt subscription because of the way
clicking "Continue" on the modal prompt subscription triggered another
subscription call. This commit makes it so that clicking "Continue"
calls a more correct and internal registration API that bypasses the
multile-subscription call check. This should have been the correct API
to use from the beginning.