Allow customization of HTTP popup text site title
Both the HTTP popup and HTTPS modal prompt can now have their Site URL
customized. For HTTP websites, the standard text displayed is the Site
URL chosen in the OneSignal dashboard Chrome platform settings. For
example, if the user's Site URL is http://site.com, then the HTTP popup
would display a large bold "http://site.com wants to show
notifications". This is now customizable to anything the user wants,
including the empty string. Also for HTTP sites, if the user chose '*'
as the Site URL, we currently display "This website", but now the user
can customize this to be anything (including the empty string).