Patch Safari violet highlight bug
Some combination of styles of the "badge" component of the notify button
is causing Safari to display a violet overlay color that seems to be
buggy behavior. Hiding the badge (using display: none) seems to fix the
overlay color. This solution can be reproduced by applying a style of
display: none; to just the badge component and by checking and
unchecking the style in the Safari DevTools inspector, the violet
overlay will hide and show accordingly.

This patch disables the badge component of the notify button on Safari.
Very few people seem to use the badge (no one has specifically asked
about it ever) and even fewer people use Safari on desktop. The badge
component on Chrome and Firefox still function normally.