Fix session bug refreshing after first subscribe
Fix bug caused by sessionStorage not setting on the host page if set in
the popup. After registering for the first time for an HTTP site, the
SDK would think this is a new session, so simply refreshing the page
would cause the session_count to increment to 2. The session_count
should always only increment on a new tab or window if the user is subscribed.