Make SDK loading compatible with CF Rocket Loader
CloudFlare's Rocket Loader rewrites DOM events in a way such that when
the DOMContentLoaded event fires (which we use to initialize the SDK),
document.readyState is still "loading", and not "interactive". We rely
on DOMContentLoaded to fire when document.readyState is "interactive",
otherwise our SDK will not be able to initialize.

This commit additionally listens to document.onreadystatechange, as
Rocket Loader still seems to fire onreadystatechange when the
document.readyState is "complete". We additionally listen for this event
to know when to load our SDK. A flag is added to prevent redundant calls
to init().