Fix notify button show/hide logic, update popover
The methods waitUntilShown() and waitUntilHidden() were not working
correctly for the notify button after we switched over to using events
from EventEmitter. These methods have now been updated and hopefully
this fixes the black circle on the notify button some users see
sometimes.

Updated popover show condition to not show if the autoRegister is false.