Reduce chance of error when displaying notifications
After a notification is displayed, a couple things happen:

- The backup notification is updated to the current notification
- The notification displayed event is broadcasted to any alive pages
- The notfication displayed webhook is executed

If any of these operations fail, a backup notification will be displayed.
However, this is not desirable and the user should not bothered by a
backup notification in these cases. This commit catches those errors and
logs them but otherwise does nothing.