Fix popup and host page post message communication
Before this commit, the popup would try to connect to the host set in
the Chrome platform settings Site URL. This would not always work
since users might enter http://site.com when their site was really at
http://www.site.com.

The popup cannot access the parent's origin since this is disallowed
for cross-origin windows.

This commit changes it so the host initates the connection to the
popup. As a minor detail, the popup has to first send a page loaded
message to the host page since the host page, being a cross origin
window, is unable to access the load event of the popup.