Remove hardcoded service worker .JS ext check when updating
ServiceWorker.ts onServiceWorkerInstalled() sets a property in IndexedDb
to remember the current service worker version. When a page referencing
the OneSignal web SDK is loaded, it checks the loaded SDK (off the CDN)
version with the remembered current service worker version. If it
doesn't match, the service worker is overwritten.

When manually testing, this hardcoded service worker .JS extension check
did not appear to block the updating of the service worker.

The effect of this hardcoded check was that, between the two service
worker versions we remember (one version for OneSignalSDKWorker, and
the second version for OneSignalSDKUpdaterWorker), the version stored
would always be for OneSignalSDKUpdaterWorker.

However, this was not a problem because our update code fortunately
says:

if (version != OneSignal._VERSION)

which means:

"If the remembered version is not equal to the SDK-loaded-off-the-CDN's
version"

then overwrite the current worker.

This compared (null !== 120080) for example, and overwrote the service
worker anyways.

But it's better if the code is written the way it should be: without a
hardcoded .JS extension check. This removes that.