Fix notification click navigating and focusing
With standard SDK settings, when a clicked notification's URL is
identical to an open tab's, that tab is focused. Any other
notification URL would open a new tab.

We offer a 'notificationClickHandlerMatch' option, which says as long as
there's an open tab on your notification URL's origin, without needing
the tab's URL to match the notification's URL exactly, we won't open a
new tab, and we'll use the existing open tab.

The bug was that this existing open tab was focused, and not navigated
to the notification's URL.

If the open tab was to https://google.com, and if the notification's URL
was to https://google.com/page, the open tab would be focused and
nothing else would happen. The correct behavior is to navigate this tab
to https://google.com/page.

One exception is if the URL is already the same. In this case, it seems
unnecessary to refresh the page. The site dev can listen to
addListenerForNotificationOpened() and hook onto the notification click
event and refresh the page if necessary.

Another exception is if client.navigate() isn't an available API.