Fix two notification click event handler issues
1) If the notification.clicked event is fired but the page is not
listening to the event, which can happen if the user never called
addListenerForNotificationOpened(), or which can happen if the user
called addListenerForNotificationOpened() once but not a second time
after an event was fired (addListenerForNotificationOpened() only works
once), then the event was originally fired without anybody receiving it
(since no listeners were installed). Now, the event is stored in the
database if there are no event listeners.

2) If a user is using notificationClickHandlerMatch: origin (special
flag), calling addListenerForNotificationOpened() retroactively on an
already clicked notification would originally NOT return any events if
the notification's URL was different from the current page's URL.
However, the entire point of notificationClickHandlerMatch: origin is to
fire the event as long as the notification's origin is the same as the
page's. This commit fixes that. This situation can really only come up
if you're using notificationClickHandlerMatch: 'origin'; any other usage
would result in opening a new window. An example of this (paired with
issue #1) is not listening to any notification.clicked event, receiving
a notification, the event is now stored in IndexedDb, you retroactively
call addListenerForNotificationOpened(), and you would now (not
originally with the bug) get the event fired.