Update notify button click listener to be more specific
The previous click listener for the 'Subscribe' and 'Unsubscribe' button
was implemented by listening to clicks on the 'window' object, and
determine whether the target was the button. This allowed the event to
be prevented from propogating if anything along the way called
event.stopPropogation().

This new method installs the click listener by using
element.addEventListener('click') directly, so that the event's source
occurs on the element itself.

This doesn't cause any compatibility issues with a previous hotfix we
recommended clients use on their website.