Fix resubscription for HTTP sites
HTTP users using the HTTP permission request would not be resubscribed
if they cleared their browser data on desktop, because their permission
would still be granted and the code did not do a complete check of
whether the user was subscribed, only a check if the permission was
granted or not.

The code has been changed to check whether the user is completely
subscribed or not. If the user is not subscribed, they are prompted to
resubscribe. The user may click "No Thanks" which follows the standard
"No Thanks" behavior.