Support custom subdomain iframe on HTTP sites
An HTTPS iFrame on an HTTP page isn't allowed to retrieve service
worker registrations, since this feature is available on secure origins
only. Our isPushNotificationsEnabled() function will try to retrieve
service worker registrations on custom subdomain environments, which
causes the error "Only secure origins are allowed". Our SDK should check
whether the current frame context's parent is an insecure origin, and if
it is an insecure origin, it should not attempt to get service worker
registrations.

This commit uses navigator.serviceWorker.getRegistrations() as a check
against whether the parent frame (if there is one) is insecure.
Normally, window.top.url isn't accessible within a cross-origin iframe,
so this function serves as our check. It will throw an error if the
combination of the child and parent origin is insecure. Normally, it
resolves with either undefined or the actual service worker
registration.