Remove closed issues from changelog
Often times, an issue will be closed after the fix is actually applied
(after the release of the SDK the issue was relevant for). The closed
issues then end up appearing in the next SDK release, or even two or
three SDK releases down the line.

This is confusing and looks like the issue was fixed in a later version
of the SDK. The "Full Changelog" link is the right thing to view to see
changes between SDK versions.