Safe initialize to empty object if null options
Just to be safe, initialize to an empty object if no options are passed.
In the few location sthat we are calling initHttp(), we always pass
options (even if the option values are blank), but in case we do not
pass an option, the SDK would not initialize since OneSignal.config
would be null and we don't want it to hang in that case.