Fix wrong alternate SW being installed
When installing or updating the service worker, we always want to
install the main (non-updater filename) worker.

This is done by checking whether the alternate worker is installed, and
if it is, we reinstall the main worker.

The previous check wasn't robust because it ignored service workers that
were in the process of installing but not yet activated. This commit
checks to make sure there isn't an installing worker before running the
same check.