Fix prompt dismiss time not increasing
This function is called in both the top frame context and the proxy
origin context. The function executes first in the proxy origin context,
and tries to increase the promptDismissCount by 1.

But because it's not awaited on, the database write doesn't complete
before the top frame grabs the existing value (still 0, because the
async operation didn't complete) and writes 0 again to the database. The
final value never increases.

This commit fixes this by waiting on the first operation.