Use the build arguments for staging environment URLs.
The current staging environment URLs rely on the `window.location.host` value.  This won't work for
application that's are being hosted on different hosts from where the SDK files are being loaded
from since the external resources are loaded. This change uses build arguments instead of the
browser environment to build those URLs.