Don't default to document title if no notification title present
Notification titles used to be required in browsers. We allowed
users to not provide a title, but if that happened, we defaulted
to the document.title property. This is no longer a required field
for notifications, and browsers handle it differently (i.e. Chrome
says 'Google Chrome' if no title is present).

This continues to use the default title value if it is provided
by the user by calling `OneSignal.setDefaultTitle`.