Switch to new slidedown event strings while keeping legacy support
After switching from the new names ("slidedown" over "popover") back to the old to revert a regression, here we are adding back the new convention while maintaining backwards compatibility.

Old event strings are now on the LEGACY_EVENTS parameter while new event strings are on the EVENTS parameter of the Slidedown object.