Fix TS Promises
Motivation: in this version of TS, the optionality of the value passed to the Promise resolve callback does not exist. For that reason, the value passed to `resolved` is required. It can still be omitted by creating a `new Promise<void>` which is what is being done here.

Read more: https://github.com/microsoft/TypeScript/pull/39817