Fix TS `super` errors
Motivation: in this version of typescript, it checks whether `super` is called before any invocation or assignment of a class member variable (e.g. via `this` keyword). Thus, we must pull the `super` call out of the switch statements to ensure it is picked up by TS and no transpile-time errors are thrown.