Default Injection Optional Chaining Text Config
Motivation: we want to handle the case in custom code integrations where the text parameter may be omitted. In this case, we should set the default ourselves. We can use optional chaining to handle this. It will return undefined for the text param value if it is omitted and the `getValueOrDefault` function will return the default.