SlidedownManager Implementation
Motivation: here we create a new manager that contains code for showing various slidedowns and mounting special containers such as tagging container or channel capture container.

We also want to track attempted slidedown displays via a slidedown queue which we store on the slidedown manager.

The slidedown manager leverages the `ContextInterface` on the `OneSignal` object