Attempted to use TS's "import type" feature
* Attempted to use the TypeScript feature import type for OneSignal.ts.
Importing OneSignal as a whole into every file that requires it ends up
bloating the ServiceWorker build as it puts all the page only code into
it as well. This is due to it bringing in all its dependencies.
* import type does not work with statics so had to make all properties
in OneSignal.ts instance based.
This however required a number of changes to create a new
OneSignalClass instance in the correct spots.
* In the end I wasn't able to get a single import type to work without
also requiring a `declare var ` in each file too.
* Stepping back what we should work to is removing the circular
dependency. That is we should never import OneSignal.ts into any files.
Dependency injection should be used instead.