Async/Await Prompting Fix
Motivation: this commit makes it so that initialization is unblocked by delayed prompts (in `InitHelper`).

However, a race condition occurrs (in testing env) where the `SlidedownManager.isSlidedownShowing` value is the same for multiple slidedowns and multiple are attempted to be mounted at the same time. This will need to be handled appropriately in unit tests.