`internalShowParticularSlidedown` Implementation
Motivation: this small refactor makes it easy to reuse code from the function `internalShowCategorySlidedown` for the new broad slidedown types: `Sms`, `Email`, and `SmsAndEmail`.

In some cases, these internal functions won't get the `options` parameter passed to them (e.g: from a public API function like `showCategorySlidedown` where we don't require the Cat Slidedown config be passed to the function). Instead we pull it from the config afterwards via `prompts`. 

`typeToPullFromConfig` is a sort of fallback to make sure we know what type of slidedown we are supposed to show (and/or config to look for) in the case `options` param is omitted.