New PromptsHelper Function to Assert Valid SMS Input
Motivation: we want to validate the field input. This is being done elsewhere: done at the web api level with the DOM element event listener `keyup` handler (see initialization of phone input library). That validation stores the validity state which is what we are accessing here. We are simply asserting that the input is valid, otherwise we're throwing an error.