PromptsManager `ALLOW_CLICK` Observer Refactor
Motivation: this commit refactors the allow observer to handle the various slidedown types. If the slidedown type is `push`, we should try registering for push. If not, we should do the corresponding action such as making a network call to update tags, set an email, or set a phone number.