Remove Retry Icon from Slidedown Error Button / State
Motivation: the Category Slidedown uses the retry button to indicate that updating the tags failed (i.e. the network request returned an error) and updating was unsuccessful.

For simple validation, we should not show a retry icon. If we need to add it for requests *related* to Web Prompts we can add it back later.