Invocation of ConfirmationToast after dismissing Channel Capture Slide
Motivation: here we are invoking the confirmation toast show function in order to display the toast, waiting 5 seconds, and then dismissing the toast. We also provide 1 second after dismissal before emitting the Slidedown CLOSED event, which will trigger the SlidedownManager to pop any consequent slidedowns from the queue.

This 1 second delay can serve as a temporary buffer between showings until we provide a separate customizable value for this number which can be passed down via config.