Feat: WebPrompts - changes to ChannelCaptureContainer logic, remote update functions, logic tweaks (#773)
* Fix validation bugs

Motivation: previously there were two issues:

1. we were allowing slidedowns with only one of the two web prompt fields (sms, email) to be accepted with empty fields
2. we were attempting to reset both error states regardless of slidedown type, leading to an error saying the validation element was not found

This commit introduces differentiation in the `resetInputErrorStates` function as well as more granularity in the `ChannelCaptureContainer` validation via new dedicate functions to check if individual fields are empty.

* Function to get sms value from input in the `ChannelCaptureContainer`

Motivation: this static function gets the sms value from the input

* New (unimplemented) functions for remote calls to update email, sms

Motivation: these functions are called from the `SlidedownManager` to update the player records with the submitted email, sms values in the Web Prompt

* Remove function `areBothInputFieldsEmpty`

Motivation: since we have the other two functions, this one is a bit redundant.

* `handleAllowClick` implementation update

Motivation: includes logic to call update helper functions to make remote calls to update the email and sms values.

More comments

* Fix Update Mode Bug

Motivation: fixes a bug where we weren't passing `isInUpdateMode` in the `spawnAutoPrompts` function

* Setter and Removers for Save and Error States

Motivation: creates a setter and remover for save and error states. This is a good clean up with logic for each in its own function. Will be easier to maintain code moving forward.

* New PromptsHelper function for push dependency

Motivation: new helper function returns whether the slidedown type is push dependent (i.e. push or category types)

Helpful to differentiate between push, category and sms, email and smsAndEmail types

Here we are using this to know whether to reset input field error states.

* Import nit