Clarifying comment regarding why autoPrompt is defaulting to true
Motivation: elsewhere in the codebase we have this setting defaulting to false. This might be confusing.

The reason is that here, we are defaulting to true iff (if and only if) the slidedown config exists in the initialization options. Think, if it was passed into the initialization, most likely the developer wants the slidedown to prompt.

In the other autoPrompt default definition, we default to false since it is used for the case where the entire slidedown config wasn't passed in at all. However, we still want it to be defined in order to support the programmatic showing of slidedowns e.g: via `showSlidedownPrompt`