`Slidedown` Update
Motivation: this commit updates the slidedown class. 

It now contains a member instance of the `ChannelCaptureContainer` so that we can access the CCC's member variables for sms and email input validity from the `SlidedownManager`

Includes some general refactoring.

There are now setter and remover functions for the save and error states.

Addition of the QUEUED popover event.