Fix unit test with awaitable event that fires after all allow click
The `handleAllowClick` function was broken up into smaller async functions that contain specific logic for the ALLOW event for each type of slidedown.

Due to this, it looks like the call stack order changed so that the unit test assertions would run before the full ALLOW click logic completes, breaking the test.

You might think this could be resolved by simply awaiting on the function that invokes `handleAllowClick` as well as the function itself. However, in PromptsManager, `handleAllowClick` is invoked in the handler of an independent event.

This commit fixes the test by introducing a special event used only in the testing environment we can await on before making the assertions.