Bug fix: switch to use `internalShowParticularSlidedown`
Motivation: because we want the programmatic showing of the push slidedown to use any custom configured prompt options, we need to switch to use `internalShowParticularSlidedown` since that function pulls any existing options from the config. Otherwise, it will use what is passed in.

Example of issue: Wordpress integration autoPrompting is done by injecting `OneSignal.showSlidedownPrompt()` onto the page's index page (as opposed to `autoPrompt: true`). Without this fix, we were seeing the auto-prompting that resulted from the auto-prompt spawning process AS WELL as the programmatic showing.

This must also be fixed in the Wordpress Plugin by simply removing the programmatic prompting. However, this change will need to wait a year or so to make sure most people are on the most recent version of the Wordpress plugin.

For now, we can fix this by ensuring that the programmatic prompting of the push slidedown checks to see if config options exist.