Make `internalShowSlidedownPrompt` private, refactor `showHttpPrompt`
Motivation: we want to prevent the use of this function in the future from outside the manager since `internalShowParticularSlidedown` should be used instead (it checks for existing config options to use).

We also refactor the OneSignal public function `showHttpPrompt` to simply call `showSlidedownPrompt` (there is no difference). We should leave for backwards compatibility, however.