ChannelManger scope of refactor notes
* Putting ChannelManager into place to handle all player updates is
requiring a large protion of code to be touched. Most of which is
shifted into new locations but this effects ordering of events due to
the nature of the changes. Some code has to be de-decoupled to
centralize some logic such as on_session as well.
* All of the above means very thorough manual testing should be done
as well as a large number of mocks and unit tests themselves need to be
updated to match the new internal API.

* The ChannelManager concept started in this PR is better than what we
have today but is shaping up to be only a marginal improvement with a
high cost to dev and testing time with high risk of breaking something.
* I have documented some alternatives in this commit which are much
more de-coupled. However these need more thought if they are a design
patterns we want to use SDK wide.

* A more incremental aprouch to adding better support for more delivery
channels such as SMS is to leave any push logic largely untouched.
Move currnet email logic into a new SecondaryChannelManager where it
should support email with the ability to drop in SMS support in a
follow up PR.
Once we have a design for a more long term internal de-coupling
strategy we can then move torwards a larger refactor.
In summary the work and risk of moving email into a new
SecondaryChannelManager give is low and is a incremental improvement.