Move reset styles to top
Motivation: if a style is tied in specificity, the tie goes to the last style set. Since the reset styles are meant to "clear" the styles in order to apply OneSignal styles uniformly, we should move the reset styles to the top so that the tie can go to the OneSignal-specific styles.