Switch functions to use synchronous DB reads
Motivation: instead of relying on the asynchronous IndexDB reads for the `isPushEnabled` and `isOptedOut` values, we switch to use the synchronous `LocalStorage` getters instead.

The reason for this is simple, asynchronous DB reads make use of the message queue which queues asynchronous calls and is outside of the event loop call stack.

As soon as the message queue is used, the entire call stack loses its designation as "user-initiated" even though it may have been. We must thus avoid using any executions that would require use of the message queue to ensure the requirement is met.