Synchronize user update functions in `SecondaryChannelSynchronizer`
Motivation: we need to synchronize player update functions since each operation should be an atomic operation.

This is particularly important when it comes to externalUserId updates since we need to be 100% sure we are updating the correct user. We cannot risk accidentally updating a user with another user's data tags for example.