OperationRepo: set up delta queue & processes
Motivation: first we subscribe via `setup()` the operation repo to the broadcasting of new deltas from the `modelRepo` model change observer and push those deltas to a new `deltaQueue` array we will process.

We then process the deltas in order (based on timestamp) and aggregate all changes into an operation. Which we will process in the operations queue.