HydratorBus
Motivation: the hydrator bus imports both the operation repo and the model repo in order to form a communication channel between them. When the operation repo broadcasts a request result as a result of a call to the static request service, they hydrator bus will hydrate the appropriate model and set the new value in the model repo.