HydratorBus Implementation
Motivation: The hydrator bus is a communication layer between the model and operation repos. It subscribes to operations outcomes from the operation repo and will hydrate them as needed to save to the model repo. The hydrator bus will then use the model repo’s public `set` function to update the intended model.

We use the `set` function in order to control whether a subsequent broadcast is made to the operation repo. Since the data in the hydrator bus just came from a successful operation, we avoid repeat calls by using the `set` function and telling it we are calling it from the hydrator bus.