Add propagation boolean to `add` functions to determine operation flow
Motivation: we need a way to hydrate entire models (previously just had way to hydrate individual properties).

Since sometimes we hydrating local models and not re-propagating, we must have an easy way to specify this.

We are adding the broadcasted change type of `ModelStoreHydrated` in the case that we don't want to propagate. This will result in the model repo halting execution after executing the local update.