Unblock model creation from waiting for `onesignalId`
Motivation: we should add the `onesignalId` to the model instance when it's available without blocking model repo execution.

We also introduce a public awaitable promise we can wait on prior to executing operations in the operation repo / executor flow. This will ensure the `onesignalId` is available when the HTTP request requires it to execute.

Fix `decode` function type errors.