Make Executors wait until the `onesignalId` is available on the model
Motivation: it is possible the models are not associated with a user yet, we should wait until the model becomes tied to a user id (`onesignalId`) prior to attempting to execute the operation

An example of this is logging in and then adding tags. When we login, we will set a temporary local user with blank data to hold any temporary changes. We will enqueue any operations as we normally would, though we will wait until the id becomes defined prior to executing it (the `onesignalId` returns from the login call)