OneSignalDeferred basic impl
This switches the public API from OneSignal to OneSignalDeferred being
deferred list of functions to run.

The new OneSignalDeferred API now only accepts a function as parameter
to push(). Also this function will provide a instance of OneSignal.

This commit also updates the example pages and confirmed it works. More
clean up will be done in future commits as part of the same PR.