Simplified shim loader
Greatly simplified the page SDK loading, this new behavior means we
don't need function stubs anymore
These are not needed as we expect everyone to use
`OneSignalDeferred.push(function(onesignal) {});`.
Direct calls without going through the new deferred API won't be
supported in the major release.

We also removed the check for a pre-defined OneSignal, and didn't
translate this into OneSignalDeferred as that isn't something a
developer would clobber by mistake.

In a follow up commit as part of this PR we will be removing all the
stub code itself and it's usages.