Update User classes to prevent duplicate users being created, refactor
Context: on web, we must wait to have at least one subscription prior to creating a user.

Problem: we need to prevent multiple user creations such as calling:
* addSms()  <-- sends user create
* addSms()  <-- sends user create (first call hasn't completed, doesn't know there is already a user)

before allowing the first user create call to complete.

Solution: the solution is to only treat the call as modifying an existing user if:
a) we are in the process of creating a user
b) already have a onesignalId

Relevant changes:
* expose `onesignalId` as a public string on `User` to be accessed synchronously from synchronous add<Subscription> calls.
* expose `isCreatingUser` boolean and perform logic when creating user so as to safeguard this from happening twice
* modify onesignal id available promise to return the `onesignalId` for convenience
* new private helper function in User.ts to copy the promise from identity model to User class instance