Make CoreDirector functions synchronous
Motivation: will help avoid race condition due to having to await on core director functions. Example: if we add a tag then login, due to the fact that we have to await on the core director functions, the the tag update will not happen in time before the login function force-flushes the delta queue.