For user login, send correct data based on identified status of prev user
Motivation: we want to make sure to send the correct data based on whether the previous user was identified or not.

If we are identifying an anonymous user, we should send the entire user data.

If we are upserting a user, we should only send the push subscription and external id.