Back up last valid user and restore if login fails instead of fetch
Motivation: if login fails, we were previously fetching based on the onesignal_id held in memory. However, this introduces an issue in the case that identity verification is turned on.

In this case, we would be unable to fetch and would enter a state of perpetually not having a onesignal_id and thus any subsequent login attempts would fail.

This commit fixes these issues by creating a backup of the user and restoring the "last valid user" in case of failure. A valid user is one with a onesignal_id