Fallback to lastKnownPushToken when getting current push sub model
Motivation: this function can fire *after* revoking the native permission (e.g: to update the notification_types). Thus, we should fall back to look for the lastKnownPushToken if there isn't one currently available. If the lastKnownPushToken doesn't correspond to the current push model in the model repo, the function would return undefined anyway.