Add Test Coverage
Motivation: add tests to make sure we are updating the push subscription models accordingly when the native permission changes. To do this, we look to see that we are updating two values: notification_types & enabled.

We update the TestEnvironmentHelper with more OneSignal globals needed to run our test.

We create a new test config param to control whether or not we initialize the user and push subscription.