fix iOS denied notificationPermissionChange
On iOS `navigator.permissions.query({ name: 'notifications' }`).onchange
does not fire due to a bug with the browser.
To account for this we are calling our
triggerNotificationPermissionChanged event every time the native
notification prompt is answered or dismissed.

This however does create duplicate events so this commit adds logic to
PermissionUtils.triggerNotificationPermissionChanged to prevent this.

Tested event now fire exactly once on iOS 16.5.
Tested both for accepting and declining.
Also confirmed Chrome on macOS still works