don't throw on user property update when no OSID
Removed throwing when we don't have a onesignal_id since this isn't
considered an error, as we will simply apply this property to the User
when it is created.

Longer term it might be a better
architecture to move the onesignal_id check that exists in
UserPropertyRequests up the stack.