remove no longer needed first page view check
This check is no longer needed as the session manager and model repo
handle these checks.

The original motivation to remove this is to address a macOS Safari
notification permission prompting issue where you can't await on I/O,
otherwise the prompting is flagged not done from an end-user
interaction. The isPushNotificationsEnabled() does disk I/O which we
removed in this commit.