Use SUBSCRIPTION_CHANGED for push id updates
OneSignal.coreDirector.getCurrentPushSubscriptionModel() was used to
attempt to listen when the PushSubscription.id becomes available
from it's internal _subscribeToPushModelChanges however this doesn't
work if there isn't a subscription yet.

This commit is a short term fix to make things more consistent by using
the SUBSCRIPTION_CHANGED event already being using for token updates.

Long term would be would be to make PushSubscriptionNamespace listen to
an event on another class in charge of adding a Subscription to a User,
so the global SUBSCRIPTION_CHANGED can be removed.
  - LoginManager isn't it either, which is why this path wasn't picked