fix getting PushSubscription by lastKnownPushToken
getCurrentPushSubscriptionModel() was attempting to use
lastKnownPushToken if the current token was undefined, however this
didn't account for pushToken changes. We now perform a check on the
current token first, then if nothing is found then do another check on
the lastKnowPushToken.

This fixes a bug where we would create a new subscription instead of
updating the exiting one when the pushToken changes. This happens when
the browser's notification permission is toggled off and then back on.

Lastly renamed, getCurrentPushSubscriptionModel to
getPushSubscriptionModel, to avoid confusion with the newly added
private function getPushSubscriptionModelByCurrentToken.